Changed less than sign to greater than
If you have noticed, my commit that was in on time was not working properly at all and was displaying events even when they were not supposed to be, and not displaying events when they were supposed to be. This was due to a single less than sign in my python, I changed it to a greater than sign and now it works! I hope I do not get counted off much for fixing such a small, but catastrophic in consequence, error after the deadline.